chore: create AlbumService, _AlbumShimmerPlaceholder, album provider
