+++ lib/app/features/gallery/providers/albums_provider.c.dart
@@ -1,0 +1,25 @@
+// SPDX-License-Identifier: ice License 1.0
+
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+import 'package:ion/app/features/gallery/data/models/album_data.c.dart';
+import 'package:ion/app/features/gallery/views/pages/media_picker_type.dart';
+import 'package:ion/app/services/media_service/album_service.dart';
+import 'package:photo_manager/photo_manager.dart';
+import 'package:riverpod_annotation/riverpod_annotation.dart';
+
+part 'albums_provider.c.g.dart';
+
+@riverpod
+AlbumService albumService(Ref ref) => AlbumService();
+
+@riverpod
+Future<List<AlbumData>> albums(Ref ref, {required MediaPickerType type}) {
+  final albumService = ref.watch(albumServiceProvider);
+  return albumService.fetchAlbums(type: type);
+}
+
+@riverpod
+Future<AssetEntity?> albumPreview(Ref ref, String albumId) async {
+  final albumService = ref.watch(albumServiceProvider);
+  return albumService.fetchFirstAssetOfAlbum(albumId);
+}
