+++ lib/app/features/gallery/views/components/albums/album_thumbnail.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/extensions/extensions.dart';
-import 'package:ion/app/features/gallery/providers/gallery_provider.c.dart';
+import 'package:ion/app/features/gallery/providers/albums_provider.c.dart';
 import 'package:ion/app/features/gallery/views/components/shimmer_loading_cell.dart';
 import 'package:photo_manager/photo_manager.dart';
 import 'package:photo_manager_image_provider/photo_manager_image_provider.dart';
@@ -20,10 +20,7 @@ class AlbumThumbnail extends ConsumerWidget {
     return previewAsync.maybeWhen(
       data: (asset) {
         if (asset == null) {
-          return ClipRRect(
-            borderRadius: BorderRadius.circular(12.0.s),
-            child: ShimmerLoadingCell(dimension: 50.0.s),
-          );
+          return const _AlbumShimmerPlaceholder();
         }
 
         return SizedBox.square(
@@ -41,10 +38,19 @@ class AlbumThumbnail extends ConsumerWidget {
           ),
         );
       },
-      orElse: () => ClipRRect(
-        borderRadius: BorderRadius.circular(12.0.s),
-        child: ShimmerLoadingCell(dimension: 50.0.s),
-      ),
+      orElse: () => const _AlbumShimmerPlaceholder(),
+    );
+  }
+}
+
+class _AlbumShimmerPlaceholder extends StatelessWidget {
+  const _AlbumShimmerPlaceholder();
+
+  @override
+  Widget build(BuildContext context) {
+    return ClipRRect(
+      borderRadius: BorderRadius.circular(12.0.s),
+      child: ShimmerLoadingCell(dimension: 50.0.s),
     );
   }
 }
