+++ lib/app/features/gallery/views/pages/album_selection_page.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/extensions/extensions.dart';
+import 'package:ion/app/features/gallery/providers/albums_provider.c.dart';
 import 'package:ion/app/features/gallery/providers/gallery_provider.c.dart';
 import 'package:ion/app/features/gallery/views/components/albums/album_item.dart';
 import 'package:ion/app/features/gallery/views/pages/media_picker_type.dart';
